+++ src/Microsoft.Azure.SignalR.Serverless.Protocols/Constants.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
+namespace Microsoft.Azure.SignalR.Serverless.Protocols
 {
     internal static class ServerlessProtocolConstants
     {
