+++ src/Microsoft.Azure.SignalR.Serverless.Protocols/ServerlessMessage.cs
@@ -3,13 +3,16 @@
 
 using Newtonsoft.Json;
 
-namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
+namespace Microsoft.Azure.SignalR.Serverless.Protocols
 {
-    internal class InvocationMessage : ISignalRServerlessMessage
+    public abstract class ServerlessMessage
     {
         [JsonProperty(PropertyName = "type")]
         public int Type { get; set; }
+    }
 
+    public class InvocationMessage : ServerlessMessage
+    {
         [JsonProperty(PropertyName = "invocationId")]
         public string InvocationId { get; set; }
 
@@ -18,5 +21,15 @@ namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
 
         [JsonProperty(PropertyName = "arguments")]
         public object[] Arguments { get; set; }
+    }
+
+    public class OpenConnectionMessage : ServerlessMessage
+    {
+    }
+
+    public class CloseConnectionMessage : ServerlessMessage
+    {
+        [JsonProperty(PropertyName = "error")]
+        public string Error { get; set; }
     }
 }
