+++ src/SignalRServiceExtension/Microsoft.Azure.WebJobs.Extensions.SignalRService.csproj
@@ -10,4 +10,12 @@
     <PackageReference Include="Microsoft.Azure.WebJobs" Version="$(MicrosoftAzureWebJobsPackageVersion)" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="TriggerBindings\Utils\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Microsoft.Azure.SignalR.Serverless.Protocols\Microsoft.Azure.SignalR.Serverless.Protocols.csproj" />
+  </ItemGroup>
+
 </Project>
