--- src/SignalRServiceExtension/TriggerBindings/Protocal/CloseConnectionMessage.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Newtonsoft.Json;
-
-namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
-{
-    internal class CloseConnectionMessage : ISignalRServerlessMessage
-    {
-        [JsonProperty(PropertyName = "type")]
-        public int Type { get; set; }
-
-        [JsonProperty(PropertyName = "error")]
-        public string Error { get; set; }
-    }
-}
