--- src/SignalRServiceExtension/TriggerBindings/Protocal/ISignalRServerlessMessage.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
-{
-    internal interface ISignalRServerlessMessage
-    {
-        int Type { get; set; }
-    }
-}
