+++ src/SignalRServiceExtension/TriggerBindings/Utils/JsonMessageParser.cs
@@ -1,62 +1,20 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Buffers;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
 
 using Microsoft.AspNetCore.SignalR.Protocol;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using Microsoft.Azure.SignalR.Serverless.Protocols;
 
 namespace Microsoft.Azure.WebJobs.Extensions.SignalRService
 {
     internal class JsonMessageParser : MessageParser
     {
-        private const string TypePropertyName = "type";
+        private static readonly IServerlessProtocol ServerlessProtocol = new JsonServerlessProtocol();
 
-        public override bool TryParseMessage(ref ReadOnlySequence<byte> buffer, out ISignalRServerlessMessage message)
-        {
-            var textReader = new JsonTextReader(new StreamReader(new ReadOnlySequenceStream(buffer)));
-            var jObject = JObject.Load(textReader);
-            if (jObject.TryGetValue(TypePropertyName, out var token))
-            {
-                var type = token.Value<int>();
-                switch (type)
-                {
-                    case ServerlessProtocolConstants.InvocationMessageType:
-                        message = SafeParseMessage<InvocationMessage>(jObject);
-                        break;
-                    case ServerlessProtocolConstants.OpenConnectionMessageType:
-                        message = SafeParseMessage<OpenConnectionMessage>(jObject);
-                        break;
-                    case ServerlessProtocolConstants.CloseConnectionMessageType:
-                        message = SafeParseMessage<CloseConnectionMessage>(jObject);
-                        break;
-                    default:
-                        message = null;
-                        break;
-                }
-                return message != null;
-            }
-            message = null;
-            return false;
-        }
+        public override bool TryParseMessage(ref ReadOnlySequence<byte> buffer, out ServerlessMessage message) =>
+            ServerlessProtocol.TryParseMessage(ref buffer, out message);
 
         public override IHubProtocol Protocol { get; } = new JsonHubProtocol();
-
-        private ISignalRServerlessMessage SafeParseMessage<T>(JObject jObject) where T : ISignalRServerlessMessage
-        {
-            try
-            {
-                return jObject.ToObject<T>();
-            }
-            catch
-            {
-                return null;
-            }
-        }
     }
 }
